feat: begun refactor of state management
The UI will become more reactive if we can subscribe to
DBus signals and update the UI based on them.

So a new channel must be created to get messages from the
signals.

The current idea is to create a SignalManager this will
recieve updates about which signals it should listen
for. These updates will have to be sent through from
the network action as they would only be known then.

SignalManager will then send an update request through
a network action.

Example:
Starting a scan makes SignalManager listen to 'scanning'
signal, when response is recieved from it we know scanning
is done. SignalManager now sends a network action to get
the current networks as well as removing the signal.