fix: moved network into com
Created a new state folder for files related
to state inside of the com package. This
currently includes signal and network.

Network was moved as signal depended on
an enum and since com doesn't depend
on tui it had to be moved.

Network does fit more naturally into
com rather than tui.